refactor: change store -> user
